+++ gvmtools/helper.py
@@ -17,12 +17,13 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import getpass
+import sys
 
 from gvm.errors import GvmError
 from gvm.xml import pretty_print
 
 
-__all__ = ['authenticate', 'pretty_print']
+__all__ = ['authenticate', 'pretty_print', 'run_script']
 
 
 def authenticate(gmp, username=None, password=None):
@@ -62,3 +63,19 @@ def authenticate(gmp, username=None, password=None):
     except GvmError as e:
         print('Could not authenticate. Please check your credentials.')
         raise e
+
+
+def run_script(path, global_vars):
+    """Loads and executes a file as a python script
+
+    Arguments:
+        path (str): Path to the script file
+        vars (dict): Variables passed as globals to the script
+    """
+    try:
+        file = open(path, 'r', newline='').read()
+    except FileNotFoundError:
+        print('Script {path} does not exist'.format(path=path), file=sys.stderr)
+        sys.exit(2)
+
+    exec(file, global_vars) # pylint: disable=exec-used